refactor(reborn): split wasm runtime module
